Created threshold and canny edge detection classes
The canny edge detection is proven to be more beneficial than the threshold edge detection
for this application. A track bar is wired to the key parameters of each of the contour
detection algotithims to find the ideal parameters for a specified image. Only the canny
edge detection algorithim has a transfer function feature which converts pixel area
into into inches and micrometers via a calibration procedure.